Fix typo in dependabot yaml (#7477)
## Summary of changes

Try to fix dependabot... again...

## Reason for change

Apparently dependabot still isn't happy:

```
The property '#/updates/0/registries' of type string did not match one or more of the required schemas
The property '#/updates/1/registries' of type string did not match one or more of the required schemas
The property '#/updates/2/registries' of type string did not match one or more of the required schemas
The property '#/updates/3/registries' of type string did not match one or more of the required schemas
The property '#/updates/4/registries' of type string did not match one or more of the required schemas
The property '#/updates/5/registries' of type string did not match one or more of the required schemas
```

## Implementation details

Make it an explicit string 🤷‍♂️ 

## Test coverage

I _wish_ I could...

## Other details

This really _Shouldn't_ fix it. but you know, YAML...


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->