Fix dependabot again by allowing private feed access (#7480)
## Summary of changes

- Allow dependabot to access the private nuget feed (hopefully)
- Remove unused files

## Reason for change

Dependabot is still broken, and we want to fix it

## Implementation details

- Delete some leftover files (that should have been removed in #7469
:oops:)
- Allow access to the private (not really, it's public) feed in
NuGet.config

## Test coverage

👀 

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->