Sign artifacts using 0.3.5 of the `dd wcs` tool (#7492)
## Summary of changes

- Update the gitlab build image to use the new dd wcs tool
- Test with the new image to make sure signing works

## Reason for change

The old version of `dd wcs` requires IMDSv1, which is deprecated. We
need to update to the new version of the tool to support v2.

## Implementation details

- Update the tool version (thanks @zacharycmontoya)
- Build and upload the windows docker image to dockerhub
- Test (once) with the dockerhub image
- [x] Switch to the mirror image once confirmed

## Test coverage

- [x] Confirm building and signing works with the new tool with the
dockerhub image
- [ ] Confirm building and signing works with the new tool with the
mirrored image

## Other details

Blocked by 
- https://github.com/DataDog/dd-trace-dotnet/pull/7490
- https://github.com/DataDog/images/pull/7647

Fixes:
- https://datadoghq.atlassian.net/browse/LANGPLAT-786
- https://datadoghq.atlassian.net/browse/LANGPLAT-787

<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>