Decrease log level when uncompressing files in Nuke (#7496)
## Summary of changes

We are getting our logs flooded by thousand of lines when uncompressing
files in our Nuke tasks. We display a line for every directory created
when uncompressing. This creates thousand of useless lines in the task
DownloadLibDatadog, which floods the output, so we stop capturing logs
in some gitlab tasks and missing some information. Since we don't need
these lines, getting ride of them would help us to get the log
information that we need.

The problem is that this is done by the Nuke library when calling the
method CompressionTasks.UncompressZip. In order to fix that, we have
created our own version of that method, which is a copy of the Nuke one
without all the unnecessary logging.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>