Fix macos build (#7503)
This PR fixes the macos build

It's failing on CI with:
```log
06:45:22 [INF] > /usr/local/bin/cmake -B /Users/runner/work/1/s/obj_arm64_x86_64 -S /Users/runner/work/1/s -DCMAKE_BUILD_TYPE=Release -DUNIVERSAL=OFF
06:45:22 [DBG] -- Configuring incomplete, errors occurred!
06:45:22 [ERR] CMake Error: CMake was unable to find a build program corresponding to "Unix Makefiles".  CMAKE_MAKE_PROGRAM is not set.  You probably need to select a different build tool.
06:45:22 [ERR] CMake Error: CMAKE_C_COMPILER not set, after EnableLanguage
06:45:22 [ERR] CMake Error: CMAKE_CXX_COMPILER not set, after EnableLanguage
06:45:22 [ERR] Target CompileNativeLoaderOsx has thrown an exception
```

Ensure the environment variables exist.

A successfully build is the test.

<!-- Fixes #{issue} -->

<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->