Simplify Windows benchmarks credential setup (#7498)
⚠️ Don't merge this PR before going through the <a
href="#checklist-for-merging-this-pr"
id="checklist-for-merging-this-pr">checklist for merging this PR</a>.

## Summary of changes

- Simplify ephemeral infrastructure credential setup on microbenchmarks
and macrobenchmarks.
- Possible due to updates on the
[tooling](https://github.com/DataDog/benchmarking-platform-tools/pull/106)
we use for managing ephemeral benchmarking infrastructure.

## Reason for change

https://datadoghq.atlassian.net/browse/APMSP-2282

## Implementation details

## Test coverage

- Successful microbenchmark run:
https://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-dotnet/-/jobs/1122255076
- Successful macrobenchmark runs:
http://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-dotnet/-/jobs/1122539541

## Other details

### Checklist for merging this PR


- [x] Merge [changes on `dd-trace-dotnet/micro` on benchmarking
](https://github.com/DataDog/benchmarking-platform/pull/193)
- [x] Revert image tag to `.*:dd-trace-dotnet-micro` on
`microbenchmarks.yml`
- [x] Revert image tag to `.*:dd-trace-dotnet-macro` on
`macrobenchmarks.yml`
- [x] Build the `.*:dd-trace-dotnet-micro` image.
- [x] Build the `.*:dd-trace-dotnet-macro` image.
- [x] Merge this PR.

<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->