Check for errors when signing (#7490)
## Summary of changes

This PR ensures that the signature of the signed files is correct and
adds some additional output.

Previously, errors in signatures were not detected because dd-wcs would
return a 0 code even when there were errors during signature. Now, we
check the output of dd-wcs and also we check later that the files are
signed, launching an exception if there were errors.

## Reason for change

We were not detecting signature errors, which could lead to publishing
unsigned artifacts.

## Implementation details

## Test coverage

## Other details

https://datadoghq.atlassian.net/browse/LANGPLAT-784
https://datadoghq.atlassian.net/browse/LANGPLAT-785

<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>