Fix incorrect usages of TracerManager PerTraceSettings (#7488)
## Summary of changes

Fix incorrect direct-usages of `PerTraceSettings`

## Reason for change

The whole point of `PerTraceSettings` is that they may change, but that
they're constant _per trace_. Therefore we should always access them
directly on the `Tracer` instance, as that insures we use the "cached"
settings on the `TraceContext` if a trace has already started.

## Implementation details

Use `Tracer.CurrentTracerSettings` instead of
`Tracer.TracerManager.PerTraceSettings`.

## Test coverage

Covered by existing, only different in some dynamic config scenarios, so
not explicitly tested (and not easy to do so)

## Other details

IMO there's a bunch of usability/naming and correctness issues:
- `CurrentTracerSettings` returns a `PerTraceSettings` instance
- `PerTraceSettings` doesn't just contain settings
- `PerTraceSettings` should contain _more_ values that can change at
runtime, basically everything changeable by dynamic config or config in
code.

Addressing those issues is something I'm trying to do soon, so this is
part of a general stack of changes coming related to config.


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->