[Dynamic Instrumentation] Added native hot standby mode to support Remote Enablement and Stable Config (#7441)
## Summary of changes
Added hot standby mode to Dynamic Instrumentation and Exception Replay
products in the native side of the tracer to support Remote Enablement
and Stable Config.

For DI/ER products to function correctly, the native side, among other
things, need to inject a new field into each and every `StateMachine`
type that is generated for async methods. Our injected field is used
during the execution of DI/ER instrumentation to capture continuation of
async flows.
The only point in time where it's possible to change type's layout is in
`ModuleLoadFinished` event, and this event is fired only once per
module.
Until Stable Configuration we only injected that field when DI/ER env
vars were enable, but now we can not rely on these env var anymore as
they're no longer the source of truth. With Stable Config we can not
know if DI/ER will be enabled or not at startup. To accommodate that,
the native side now acts as if DI/ER is enabled, just in case it will.
This is true not only for Stable Config, but also for Remote Enablement.

There are two new env vars:
1. `DD_DYNAMIC_INSTRUMENTATION_MANAGED_ACTIVATION_ENABLED`
2. `DD_EXCEPTION_REPLAY_MANAGED_ACTIVATION_ENABLED`

These env vars, as described in [Stable Config Standby mode
doc](https://docs.google.com/document/d/1KEu4rgOnB3VOTiB_uztJGwT6w4dQ6FNCryVnMQItiP8/edit?tab=t.6njjxcv5o5he#heading=h.hgsgqqfp3ubt),
are used to deactivate hot standby for ER/DI.

A big contribution to support remote enablement for DI/ER involved
changing our injection strategy of async methods.
In DI/ER instrumentation, on `ModuleLoadFinished`, we inject a field of
type `AsyncDebuggerState` into all `StateMachine` types. This field is
defined in `Datadog.Trace`. When the managed loader fails/bails out in
loading `Datadog.Trace`, it results in TypeLoadException as the runtime
tried to implicitly load `Datadog.Trace` to no avail.
In this PR, the field type is changed to be `System.Object` to avoid
that cumbersome.

## Implementation
The field that is injected and used in async methods to determine
continuation is changed to be `System.Object`. To support that, the
instrumentation of async methods for log probes has been changed -
instead of accepting `AsyncDebuggerState` as the state, it's now
accepting `System.Object`.
The entrypoints now goes through a new shim class
`AsyncMethodDebuggerInvokerV2`, and from there to the underlying
`AsyncMethodDebuggerInvoker` after materializing the state from
`System.Object` -> `AsyncDebuggerState`.

In span probes on async methods the same concept has been applied, but,
instead of a new class shim, new signatures were added that accept the
state as `System.Object` and in similar fashion materializing to
`AsyncDebuggerState` before transferring control the the previous
functions.

## Test
Relying on existing test suite of debugger & tracer to cover the
instrumentation changes

## Reason for change
Fully support Remote Enablement and Stable Config.

Resolves DEBUG-4421