Fix `ManualInstrumentationConfigSource` and introduce base type (#7508)
## Summary of changes

- Fixes a bug when using newer version of auto tracer with <3.7.0 of
manual package
- Introduce a shared config source base type

## Reason for change

While working on other config stuff, I noticed that we were accidentally
never using the "legacy" config source that was introduced in #6397.
This source would only be used if the customer is using < 3.7.0 of the
NuGet package with >= 3.7.0 of auto instrumentation, but it would be
"bad" if they did 😅

Additionally, for other other upcoming changes it made sense to create a
dedicated common base class, and expose the `useDefaultSources` flag.

## Implementation details

- Introduce a base class
- Expose `UseDefaultSources` property on base class
- Make sure we use the legacy source when we're supposed to

## Test coverage

:awkward-monkey:

## Other details

Part of a config stack

https://datadoghq.atlassian.net/browse/LANGPLAT-819


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->