Allow passing fallback values to `IntegrationSettingsCollection` (#7510)
## Summary of changes

Allow passing an existing `IntegrationSettingsCollection` in to use as
"fallback" values

## Reason for change

Needed functionality as part of a general config refactor (coming in
subsequent PRs). We need to be able to find a configuration source to
look in, but if the values aren't there, use the existing collection.

## Implementation details

Allow optionally passing in an `IntegrationSettingsCollection` when
creating a new `IntegrationSettingsCollection` to use as the fallback
values

## Test coverage

Added a unit test for the behaviour, but it's otherwise not used yet.

## Other details

https://datadoghq.atlassian.net/browse/LANGPLAT-819


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->