Upload execution benchmark log files and cleanup (#7521)
## Summary of changes

- Fix error trying to delete log file in use
- Upload execution benchmark log files for the job
- Make failures more obvious

## Reason for change

We want to try to understand spurious results, don't want to be confused
by unrelated messages, and want to make it easier to understand failurs

## Implementation details

Tweak some scripts

## Test coverage

I'll check that this run is doing what we expect

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->