Hotfix v3.26.3 - Update CI with Windows signing remediations (#7527)
## Summary of changes

This cherry-picks the commits related to resolving the issues that we
had with not correctly signing Windows artifacts.

## Reason for change

The remediations were only on `master` causing the hotfix to not be
signed again. This resolves that.

## Implementation details

`git cherry-pick` the three commits

## Test coverage

If the build works and says that everything gets correctly signed again
then 👍

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>
Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>
Co-authored-by: NachoEchevarria <53266532+NachoEchevarria@users.noreply.github.com>
Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>