fix: change Language value from '.NET' to 'dotnet' (#7523)
## Summary of changes

Change the lang name to same as managed side.

## Reason for change

Telemetry events are not ingested.

## Implementation details

### Before

Managed
```json
      [
        "Dd-Client-Library-Language",
        "dotnet"
      ],
   .....
    "length": 24733,
    "content": {
      "api_version": "v2",
      "naming_schema_version": "0",
      "request_type": "message-batch",
      "tracer_time": 1758012828,
      "runtime_id": "aaef7bef-24b0-4707-8a56-2034c3b31eed",
      "seq_id": 1,
      "application": {
        "service_name": "weblog",
        "env": "system-tests",
        "service_version": "1.0.0",
        "tracer_version": "3.27.0.0",
        "language_name": "dotnet",
        "language_version": "8.0.20",
        "runtime_name": ".NET",
        "runtime_version": "8.0.20"
      },
```

libdatadog
```json
  [
        "Dd-Client-Library-Language",
        ".NET"
      ],
....
    ],
    "length": 599,
    "content": {
      "api_version": "v2",
      "tracer_time": 1758012829,
      "runtime_id": "aaef7bef-24b0-4707-8a56-2034c3b31eed",
      "seq_id": 1,
      "application": {
        "service_name": "weblog",
        "language_name": ".NET",
        "language_version": "8.0.20",
        "tracer_version": "3.27.0.0"
      },
```

### After

Waiting for system tests to finish.



With this change, we are going to have same language name in libdatadog
as managed ie `dotnet`.

## Test coverage

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>