Hide problematic public APIs using `[EditorBrowsable]` (#7531)
## Summary of changes

Hides some of the public APIs exposed in Datadog.Trace that we don't
think customer should be using.

## Reason for change

There are several configuration APIs that customers can call using
configuration in code which are problematic:

- `new TracerSettings()` - ignores all env vars and default sources.
Customers should generally be calling
`TracerSettings.FromDefaultSources()`, it's unlikely that they really
_want_ this behavior. It also makes our code much more complex.
- `new TracerSettings(bool useDefaultSources)` - as above
- `TracerSettings.set_StatsComputationEnabled` - changing this at
runtime is problematic, particularly with the move to libdatadog data
pipeline, and at some point soon we'll basically be ignoring it
- `TracerSettings.set_AgentUri` - changing this at runtime means
everything has to reinitialialize, and customers likely end up with
incorrect values for some period of time, as well as increased startup
times
- `TracerSettings.Exporter.set_AgentUri` - as above

We would _like_ to obsolete and remove these properties, but that takes
two major versions to do (one to obsolete them, one to remove them).
This at least ensures that _new_ customers shouldn't use these APIs.

## Implementation details

Add `[Browsable(false)]` and
`[EditorBrowsable(EditorBrowsableState.Never)]` to the APIs.

## Test coverage

There's no change in functionality, so nothing really to test. Updated
the public APIs. This requires an additional in-box reference for .NET
Core 3.1 + .NET 6, but I think that's fine.

## Other details

https://datadoghq.atlassian.net/browse/LANGPLAT-819