Add a Continuous Profiler availability check (#7534)
## Summary of changes

Add a checker to decide whether the profiler is available

## Reason for change

We are going to be (and are currently) making P/Invoke calls on startup
to call the profiler. If the ClrProfiler is not attached, or the CP file
is not available (and so P/Invokes are not re-written) then this can
cause additional delays.

## Implementation details

Add explicit checks for
- Unsupported environments
- Has the profiler set the "profiler available" env var (only works on
Windows right now)
- Are we in known unsupported environments (lambda and Azure functions
outside of AAS extension)
- Is the ClrProfiler attached

## Test coverage

If everything still works, this _should_ work

## Other details

A blocker for #7287