Fix many ad-hoc grabbing of settings objects
This "fixes" the places that were previously grabbing the environment/service etc from TracerSettings to use CurrentTracerSettings instead.

This is effectively still just a refactoring, but prepares for the point where these settings aren't replaced on TracerSettings. The updates in this PR are for cases where you don't have long-lived services, and rather need to do ad-hoc Tracer.Instance grabbing of the setting values in a global context. Note too that many of these places can be updated in the future to subscribe to changes if that provides performance benefits. Also note that I elected not to change must calls to `IsIntegrationEnabled` in this PR as there are hundreds of locations,

Also found a few cases that were incorrectly assuming that these values cannot change. Marked them with 'TODO: Subscribe to changes in settings'