Remove the guards around M.D.SqlClient V4 in tests (#7541)
## Summary of changes

there was a piece of code skipping some tests in v4 and above of the
library. Those were added in #2077 referencing
https://github.com/dotnet/SqlClient/issues/1390
However, this check has stayed in place ever since, stopping tests from
being run on newer versions, even though the issue says that the issue
was fixed in 4.0.1 if I read correctly. Since we don't run tests on
v4.0.0 (the lowest 4 we run on is 4.0.6), we shouldn't need to check for
this anymore.
This might have been kept around because of an other issue where the
sample app encountered an authentication issue trying to connect to the
server in docker because the security certificate was not proper. Adding
`TrustServerCertificate` on the connection string fixes this, and should
allow us to run all tests always.

## Reason for change

better test coverage for SqlClient

## Implementation details

## Test coverage

ran a pipeline testing all package versions for this test here:
https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=187543

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->