Remove duplicate "public" APIs and `Internal` prefix/suffixes (#7546)
## Summary of changes

- Removes APIs marked with `[Public]` which are no longer used
- Remove `Internal` suffixes and prefixes

## Reason for change

The `[Public]` attributes were introduced in v2 when we needed to
segregate APIs customer's invoke from our own internal APIs. To work
around that, we created "duplicate" APIs for many functions, and
decorated the public ones with `[Public]`. However, in v3, we have a
separate shim, so this is no longer necessary, and those public APIs are
entirely unused (outside of testing).

## Implementation details

- Remove unused APIs
- Where `Internal` versions of the APIs exist, rename them and update
calling code
- Left _some_ `[Public]` attributes for methods that we still don't
really want to call from inside Datadog.Trace
- Will clean up this in a separate PR (e.g. rename `[Public]` to
`[TestingOnly]`) to give the same usage guards but being more explicit
about the purpose

## Test coverage

Just removing dead code, so as long as the tests still pass, we should
be good

## Other details

Partly related to https://datadoghq.atlassian.net/browse/LANGPLAT-819
and the config stack, as want to avoid needing to handle unused APIs