Simplify and fix reloading issue
Whenever you open up a code file where the error
was the error would disappear.

This was because the syntax tree changes when
you open it and Roslyn appears to be swapping
the type to "object" and boxing it.
I noticed this earlier but thought I resolved it.

Ended up going for a much simpler approach that
seems consistent.