Update the `ProfilerSettings` enablement to ignore SSI (#7557)
## Summary of changes

Updates the rules around `ProfilerSettings` enablement

## Reason for change

After #7287, we decide whether the profiler is enabled in managed code,
instead of on the native side. However, we got some of the rules wrong.
Specifically, we believed that the profiler should behave different when
`DD_INJECTION_ENABLED` was set, but that's not the case. The profiler
should behave the same in all cases.

## Implementation details

Fixed the implementation to ignore the SSI variable, and updated the
corresponding unit tests

## Test coverage

Updated the tests that were affected

## Other details

The incorrect behaviour was actually introduced in #7303, but it didn't
have a visible impact until #7287