Update Datadog.Sketches vendored code (#7299)
## Summary of changes

Re-vendor the
[Datadog.Sketches](https://github.com/DataDog/sketches-dotnet) code.

## Reason for change

We'd like the changes from @brontosaurusgym in
https://github.com/DataDog/sketches-dotnet/pull/18 which fixed a bug
with `MergeWith()` calls.

## Implementation details

I opted to not cut a new release for Datadog.Sketches and just pointed
it to grab from `main` instead to update the code.

~~I then undid the `UpdateVendoredCode` changes and re-ran to just get
the correct URL back which means that if we do cut a new release it will
work normally, but if we don't cut a release it won't make changes.~~

Main changes are the `store.` additions, it appears the other files
haven't been run since `UpdateVendoredCode` has been changed.

## Test coverage

## Other details

Thanks @brontosaurusgym

<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->