Add a code fix provider for invalid DuckType null checks (#7448)
## Summary of changes

This introduces a new Roslyn analyzer to flag instances where we are
directly checking if an `IDuckType` is or isn't `null`.
In most cases this is incorrect and the correct equality check would be
to check if the `IDuckType.Instance` is or isn't `null`.

The analyzer flags all instances of the former as a new rule called
`DDDUCK001` 🦆 .

In addition to the analyzer, I've added a code fix provider that
rewrites the flagged expression to use `?.Instance` instead. This allows
us to correct the issue _and_ still check for `null` on the base
`IDuckType` just to be safe.

**It is important to note though that not every instance of this
detection is 100% accurate - there are and will be cases where checking
if the `IDuckType` is `null` is the correct thing to do.**

However, using the conditional accessor should still be safe, but for
transparency I've added some current cases to a `pramga` and to
`tracer/GlobalSuppressions.cs`.

To ensure that we do not potentially disrupt developer workflows
drastically with this the analyzer is flagged as a Warning, it will be
bumped to an Error at a later point.

## Reason for change

Help enforce us to not accidentally check the IDuckType for `null` and
instead check the `Instance` for `null`.

Using a null conditional (`?.`) also means that we are _still_ checking
if the `IDuckType` is `null` on the off chance that it is.

## Implementation details

The analyzer is based off of our other analyzers (somewhat) and largely
was based on stepping through the code in the tests in debug mode (works
well in VS).

It will match on the following expressions: `==`, `!=`, `is`, `is not`.
This is only when the left and right hand sides implement an `IDuckType`
and the other side is the `null` keyword.

I went through each of the current violations of the analyzer and either
applied the code fix in VS automatically or ignored them on a
case-by-case basis.

## Test coverage

- Tests have been added for the following usage patterns (code fix and
analyzer included)
- binary and `is` patterns: `duckType != null` / `duckType is not null`
  - generic constraints: where T : IDuckType
  - `(object)duckType == null` -> `duckType?.Instance == null`

## Other details
<!-- Fixes #{issue} -->

There was some ChatGPT usage here around the parenthesis / casting and
namespace exclusions as that really tripped me up.

<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->