No longer forgive `spans.Array` from being `null` (#6878)
## Summary of changes

Remove the `!` to be a `?`. I _think_ the `!` isn't accurate as `Array`
can be `null`

## Reason for change

I noticed a error in error tracking for a NullReferenceException in this
method and this looked pretty suspect to me, but we do this in other
places as well 🤔

## Implementation details

! -> ?

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->