Escape all paths in DownloadLibDatadog (#7583)
## Summary of changes

This wraps all paths in `DownloadLibDatadog` so that we don't run into
issues of spaces in the paths.

## Reason for change

Fails to build if you are running the build script from a directory that
has a space in it.

## Implementation details

Extracted some of the paths and then quoted them all.

## Test coverage

Ran locally in a directory -> `C:/Delete Me/dd-trace-dotnet/` and it
failed.
With this it passed.

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->